--- react-fm/src/pages/Match-queues/index.ts
@@ -1,3 +1,0 @@
-import MatchQueue from './queue';
-
-export default MatchQueue;
