+++ react-fm/src/pages/matchQueues/eventsComponents/Cities.module.scss
@@ -1,0 +1,10 @@
+.citiesName {
+  color: rgb(76, 233, 90);
+  font-size: 32px;
+  margin-bottom: 15px;
+  font-weight: 600;
+}
+  
+.citiesName img {
+  height: 40px;
+}
