+++ react-fm/src/pages/matchQueues/eventsComponents/Players.module.scss
@@ -1,0 +1,14 @@
+.avatar {
+  margin-right: 15px;
+}
+  
+.grid {
+  display: flex;
+  align-items: center;
+}
+  
+.playerNames {
+  color: #fff;
+  font-size: 16px;
+  font-weight: bold;
+}
