+++ react-fm/src/pages/matchQueues/eventsComponents/Players.tsx
@@ -3,7 +3,7 @@ import Avatar from '@mui/material/Avatar';
 import Grid from '@mui/material/Grid';
 import Typography from '@mui/material/Typography';
 import { avatars } from '../constants';
-import styles from '../queue.module.scss';
+import styles from './Players.module.scss';
 
 interface Props {
   displayName: string;
@@ -11,7 +11,7 @@ interface Props {
 }
 
 export const PlayerDetails: FC<Props> = ({ displayName, index }) => (
-    <Grid item xs={2} sm={6} md={4} className={styles.grid} key={index} >
+    <Grid item xs={2} sm={4} md={4} className={styles.grid} key={index} >
       <Avatar className={styles.avatar} alt="profile" src={avatars[index]} />
       <Typography className={styles.playerNames}>{displayName}</Typography>
     </Grid>
