+++ react-fm/src/pages/matchQueues/index.ts
@@ -1,0 +1,3 @@
+import MatchQueue from './Queue';
+
+export default MatchQueue;
